increasing l when encoding spherical harmonics decreases error

+++ constants.py
@@ -1,6 +1,6 @@
 import jax.numpy as jnp
 
-default_dtype = jnp.float64
+default_dtype = jnp.float32
 EVEN_PARITY = 1 # a mnemonic to remember this is: "multiplying by 1 doesn't change the sign"
 ODD_PARITY = -1
